Refactored components/Chess
