+++ src/features/Chess.js
@@ -2,24 +2,24 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { CssBaseline, Grid } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import InfoAlert from '../features/alert/InfoAlert.js';
-import OpeningSearchEcoDialog from '../features/dialog/OpeningSearchEcoDialog';
-import OpeningSearchMovetextDialog from '../features/dialog/OpeningSearchMovetextDialog';
-import OpeningSearchNameDialog from '../features/dialog/OpeningSearchNameDialog';
-import CreateInviteCodeDialog from '../features/dialog/CreateInviteCodeDialog';
-import EnterInviteCodeDialog from '../features/dialog/EnterInviteCodeDialog';
-import ProgressDialog from '../features/dialog/ProgressDialog';
-import LoadFenDialog from '../features/dialog/LoadFenDialog';
-import LoadPgnDialog from '../features/dialog/LoadPgnDialog';
-import PlayLikeGrandmasterDialog from '../features/dialog/PlayLikeGrandmasterDialog';
-import PlayOnlineDialog from '../features/dialog/PlayOnlineDialog';
-import WatchDialog from '../features/dialog/WatchDialog';
-import OpeningAnalysisTable from '../features/table/OpeningAnalysisTable.js';
-import GameTable from '../features/table/GameTable.js';
-import Board from '../features/Board.js';
+import InfoAlert from './alert/InfoAlert.js';
+import CreateInviteCodeDialog from './dialog/CreateInviteCodeDialog';
+import EnterInviteCodeDialog from './dialog/EnterInviteCodeDialog';
+import OpeningSearchEcoDialog from './dialog/OpeningSearchEcoDialog';
+import OpeningSearchMovetextDialog from './dialog/OpeningSearchMovetextDialog';
+import OpeningSearchNameDialog from './dialog/OpeningSearchNameDialog';
+import LoadFenDialog from './dialog/LoadFenDialog';
+import LoadPgnDialog from './dialog/LoadPgnDialog';
+import PlayLikeGrandmasterDialog from './dialog/PlayLikeGrandmasterDialog';
+import PlayOnlineDialog from './dialog/PlayOnlineDialog';
+import ProgressDialog from './dialog/ProgressDialog';
+import WatchDialog from './dialog/WatchDialog';
+import OpeningAnalysisTable from './table/OpeningAnalysisTable.js';
+import GameTable from './table/GameTable.js';
+import Board from './Board.js';
 import Game from './Game.js';
-import HeuristicsBar from '../features/HeuristicsBar.js';
-import MainButtons from '../features/MainButtons.js';
+import HeuristicsBar from './HeuristicsBar.js';
+import MainButtons from './MainButtons.js';
 import SupportButtons from './SupportButtons.js';
 import '../index.css';
 import store from '../store';
